added checks for unspecified generic params
