+++ axum-macros/tests/debug_handler/fail/generics_no_with.rs
