+++ axum-macros/tests/debug_handler/fail/generics_no_with.stderr
@@ -1,0 +1,5 @@
+error: Generic param is missing a specialization in `#[axum_macros::debug_handler]`. Specify each generic param at least once using the 'with' attribute to support debugging generic functions.
+ --> tests/debug_handler/fail/generics_no_with.rs:4:18
+  |
+4 | async fn handler<T>(extract: T) {}
+  |                  ^
