+++ axum-macros/tests/debug_handler/fail/generics_unspecified_param.stderr
@@ -1,0 +1,5 @@
+error: Generic param is missing a specialization in `#[axum_macros::debug_handler]`. Specify each generic param at least once using the 'with' attribute to support debugging generic functions.
+ --> tests/debug_handler/fail/generics_unspecified_param.rs:6:21
+  |
+6 | async fn handler<T, U>(_extract_t: Path<T>, _u: Json<U>) -> impl IntoResponse
+  |                     ^
